[FE] FIX: mapInfo폴더의 컴포넌트 파일 분리 #805
